OpiServer/Server_final12/Server_final/requirements.txt
absl-py==2.1.0
appdirs==1.4.4
astroid==2.9.3
asttokens==2.0.5
attrs==23.2.0
bcrypt==4.1.2
beautifulsoup4==4.12.3
blinker==1.4
boto3==1.34.65
botocore==1.34.65
build==1.1.1
certifi==2024.2.2
cffi==1.16.0
chardet==4.0.0
charset-normalizer==3.3.2
click==8.0.3
colorama==0.4.4
coloredlogs==15.0.1
configobj==5.0.6
contourpy==1.2.0
cryptography==3.4.8
cupshelpers==1.0
cycler==0.12.1
Cython==0.29.33
dbus-python==1.2.18
defer==1.0.6
distlib==0.3.8
distro==1.7.0
dlib==19.24.2
docutils==0.17.1
face-recognition==1.3.0
face-recognition-models==0.3.0
filelock==3.13.1
flatbuffers==24.3.7
fonttools==4.50.0
grpcio==1.46.3
grpcio-tools==1.46.3
gTTS==2.5.1
httplib2==0.20.2
humanfriendly==10.0
idna==3.6
importlib-metadata==4.6.4
imutils==0.5.4
iotop==0.6
isort==5.6.4
jedi==0.18.0
jeepney==0.7.1
Jinja2==3.1.3
jmespath==1.0.1
joblib==1.3.2
keyring==23.5.0
kiwisolver==1.4.5
language-selector==0.1
launchpadlib==1.10.16
lazr.restfulclient==0.14.4
lazr.uri==1.0.6
logilab-common==1.8.2
louis==3.20.0
MarkupSafe==2.1.5
matplotlib==3.8.3
mccabe==0.6.1
mediapipe==0.10.9
more-itertools==8.10.0
mpmath==1.3.0
mypy==0.942
mypy-extensions==0.4.3
netifaces==0.11.0
numpy==1.26.4
oauthlib==3.2.0
onnxruntime==1.17.1
opencv-contrib-python==4.9.0.80
opencv-python-headless==4.9.0.80
opencv-python==4.9.0.80
packaging==24.0
paramiko==3.4.0
parso==0.8.1
pexpect==4.8.0
Pillow==9.0.1
platformdirs==4.2.0
protobuf==3.19.6
psutil==5.9.0
ptyprocess==0.7.0
pycairo==1.20.1
pycparser==2.21
pycups==2.0.1
pydub==0.25.1
Pygments==2.17.2
PyGObject==3.42.1
PyJWT==2.3.0
pylint==2.12.2
PyNaCl==1.5.0
pyparsing==2.4.7
pyproject_hooks==1.0.0
PyQt5==5.15.6
PyQt5-sip==12.9.1
pyserial==3.5
python-dateutil==2.8.1
python-debian==0.1.43+ubuntu1.1
python-vlc==3.0.20123
python-xapp==2.2.1
pyttsx3==2.90
pyxdg==0.27
PyYAML==5.4.1
reportlab==3.6.8
requests==2.31.0
roman==3.3
s3transfer==0.10.1
scikit-learn==1.4.1.post1
scipy==1.12.0
SecretStorage==3.3.1
segno==1.6.1
Send2Trash==1.8.1b0
setproctitle==1.2.2
sh==1.14.3
six==1.16.0
sounddevice==0.4.6
soupsieve==2.5
speechkit==1.2.2
SpeechRecognition==3.10.1
sympy==1.12
terminator==2.1.1
thonny==3.3.14
threadpoolctl==3.3.0
toml==0.10.2
tomli==2.0.1
tqdm==4.66.2
transliterate==1.10.2
typed-ast==1.4.3
typing-extensions==3.10.0.2
ubuntu-drivers-common==0.0.0
unattended-upgrades==0.1
urllib3==1.26.5
virtualenv==20.25.1
vosk-tts==0.3.54
wadllib==1.3.6
wikipedia==1.4.0
wrapt==1.13.3
xdg==5
xkit==0.0.0
zipp==1.0.0
pyaudio==0.2.14
